Make Slide extend Serializable and add a serialVersionUID
I intend to transmit Slide instances in binary form in the finished app.
This would allow a user to run this system locally or on a client-server
basis. Slides would be transmitted between the admin GUI and the
projection display as members of a not-yet-implemented Command class so
that they could be generated by the GUI and then incorporated into the
display unit's projection stream; to do this requires that the Slides be
serializable.

The serialVersionUID was obtained from the JDK SE 7 command serialver.