Fix a stupid bug in checkTies
Specifically, I ran a for loop but was using the array being looped
through instead of the subarray in the current iteration. This meant
that even when a subarray was supposed to be null, I was looking at the
overall array, saying "this isn't null!" and returning false.

This is why JUnit rocks, guys!