Added Command
This class transmits a Message.

This whole Message/Command scheme is probably relying on reflection in
an overly shaky manner, but it's the only way I can think of to allow
the transmission of variable-type parameters without creating subclasses
of Command and forsaking Message altogether. Maybe that's the better
option? I'll have to ponder it...these commits might be reverted
tomorrow, so enjoy them while they last!