Create a skeletal Display class (will become the client)
The current implementation is bare-bones, with only a bunch of TODO
comments and a constructor that initializes some I/O streams. Hopefully,
this will change soon... really, I just need to setup a JPanel in an
undecorated JFrame and add a game-loop-esque system with command
handling. (I've always envisioned this app's display logic as being
based on the programming of a basic game, except that input would come
from the server with Command objects, not the keyboard and mouse.)