Designed ConnectionDetailsFrame (not implemented)
a.k.a. how to give your project a compiler warning

a.k.a. my first ever Swing GUI

a.k.a. a pretty derpy WindowBuilder project

This gathers connection details (including whether to use a local or
remote instance) and, in theory, sends the details back to the client.
It will probably be used by both client and server to gather connection
details.

I intend to add some input validation to the IP and port, as well as
allow the Strings displayed to be changed by a constructor. I'll also
probably allow an application to optionally disable the Submit button if
Local is selected (useful for the client, which will be instantiated by
the server in a localhost setup).

The non-.java files added by this commit are Eclipse WindowBuilder data.