Implement DisplayPanel with preliminary logic
This logic is based on a game loop, for a very good reason: I'll be
processing input (from streams instead of keyboard and mouse), rendering
elements (Slides and ticker), and placing the rendering on the screen.

This implementation is very similar to that in Killer Game Programming
in Java. See source code for ISBN. I've omitted some elements (like the
updates-per-second compensation portion; the updater will pull all of
the Commands from the stream, so just once is sufficient).