Add note to add keep-alive system to all protocol messages
The client and server must be kept in a consistent state. In order to
achieve this, I MUST demand a client response to every server command.
(This might not be in the form of specialized messages, but I do need to
confirm client receipt of every single one of my commands and notify the
administrator if the connection drops.)

With the increasing prevalence of WiFi, combined with the probability
that this system would be used in an actual client-server installation
of ResultsWizard2, this is a pressing concern for the robustness of our
application.