Add Message.XMIT_ERROR_RESTART support to StreamQueueProxy
The anonymous Callable will now return a Command implementation with a
message type of XMIT_ERROR_RESTART instead of returning null. (The
ConcurrentLinkedQueue would choke on a null value, so this is a
preferable alternative.) Clients receiving X_E_R in the input queue
should discard all I/O mechanisms except the Socket and raw binary
streams and reinitialize the object streams, StreamQueueProxy, and
queues.