Create the Fraction class
I want to have very high precision when handling sweepstakes points.
Since some mechanisms use averages to handle sweeps in situations where
competitors are tied, this could lead to rational non-integer points.
So, I will use the Fraction class for all sweepstakes point arithmetic.

This object will represent a proper fraction; it is intended to be
paired with an int that represents the unit portion. It will support
additions that would create an improper fraction, and provides methods
to retrieve the unit portion. It will also work properly when using
multiple additions in a row, even when this would lead to an improper
fraction.