Add note to extend Slide
Yes, this is network code. Slides are being sent over the network, and
this will help tremendously in case the WiFi drops and the server and
client need to reconcile state; the server won't (hopefully) have to
wipe the client and send everything all over again.

As a side note, implement-slide will probably be a very lonely branch,
since that kinda depends on this branch. Sorry, bud.