Add Fraction tests: add(Fraction)
subtract(Fraction) is not tested because it delegates to
add(Fraction.negative()). testSubtractFraction has been deleted for this
reason.

These tests should prove to be exhaustive. I might go ahead and add some
tests for perfect units, etc., as well as add some cases where
denominators are not equal but not relatively prime (i.e. gcd(denom1,
denom2) != 1), but other than that I'm done with FractionTest.