assignPoints: Convert most direct spec[] indexes to intIndexOrZero
This fixes a bug I noticed where if there are too many places in
quantities, it could overrun spec with too-high indices.

Note that the averaging routines check the indices religiously, so I
didn't make the change there.

On a side note, regex is awesome.
spec\[([^\]])+\] -> intIndexOrZero(spec, $1)

(So are backslashes, I suppose.)