Fix AVERAGE_ADJUSTED for ties of 2 or less participants
Finding this bug and implementing this fix made me realize that
AVERAGE_ADJUSTED is probably an untenable scheme in an actual
competition ruleset, but I'm not going to delete code without a good
reason. There might be some esoteric league I've never heard of that
uses this scheme, or I might be able to modify the code somewhat to fit
a slightly different scheme that deals with this problem in a different
way.