BuildableStackedSlide: Change names for commit scheme
Now, the scheme revolves around what I'd say is a typical setup for
Git--index, head, and remote. Only what is on the remote is rendered;
the other two serve as undo buffers that can be either cleared or pushed
up to the next level at any time.

I'm not claiming that this class is a full version control system, but
hey--I'm a programmer and this is back-end non-luser-facing code, so I
can't resist using familiar names for things...