Add DERSL.forceAddTie
This forces the addition, using multiple slides if needed. It also
pushes, unlike tryAddTie, which only commits (it's waiting for the
entire block of results, or entire special honor, to be authored before
pushing). This is because if we're forcing a tie onto the slide, we
already know we want it there without a question, so there's no point in
reverting it.