Process student names prior to display by truncating at the ` mark
The admin GUI will enforce unique student names per event (this avoids
messing up the sweepstakes logic, which will erroneously add the points
of two students with the same name together). To get around this, if two
John Smiths compete in the same tournament, an admin can add text after
a grave accent to each name--for example, John Smith`TEXT. Since
students are not tracked outside of their events (only schools, for
sweeps computation), an admin need not keep track of this across events.
(We don't do this for schools--admins will probably want to ensure that
school names are unique, because two different Franklin High Schools can
be pretty confusing to viewers, while two different John Smiths aren't a
problem if they attend different schools. Besides, if a crazy programmer
founds echo `date` High School, we can handle it here...hehe.)

We also trim student names. This is for several reasons. First, if a
user prepends or appends spaces to someone's name (which I expect people
to do as a uniqueness solution, even though graves are prescribed), it
won't shrink the name rendering inordinately. Second, it ensures that
people's names always line up and aren't sized inordinately small
(because spaces, while invisible, do take up room when rendered and
could affect the fit-to-size logic). Third, it ensures that if a user
puts a space between a competitor's name and a grave--for example, John
Smith `TEXT--that the extra space won't show up on the projection.