Remove unused Graphics parameter to DERSL constructor
There is literally no reason why this thing exists. All the Slides
adjust themselves to the given Graphics on the fly, so we don't need to
pass in a g parameter to the Slides until we call draw.

Since the g parameter to the DERSL constructor is misleading, I thus
removed it.