Wipe SweepstakesAssignment.CUSTOM off the map [REASON INSIDE]
WHY DID I REMOVE CUSTOM?
========================

Originally, my plan for SweepstakesAssignment was for its assignments to
be "set in stone." There would be no way to change what you got from the
assignPoints method. To rectify this, I intended for sweepstakes points
to be user-editable if given a special configuration; this ended up
being SweepstakesAssignment.CUSTOM ("c" in the event config).

Later, I realized that it would probably be better if all sweeps results
were modifiable by the user before publication. The
SweepstakesAssignment would serve as a sensible default that the user
could use if it produced correct results. If not, the user could make
edits, if needed. (This also will serve to help mitigate the effect of
bugs in my logic, though I like to think that my unit tests will have
caught all of those.)

My intent is to add methods to Sweepstakes to make the sweeps totals
editable. Totals will be editable per competitor and passed in as if by
the results of assignPoints: as a "condensed" int[][]. I will also need
to ensure that DERSL's addTie methods doesn't assume that all
competitors in a tie have the same amount of points; this will require a
signature change for those methods.