Add ArrayUtils.expandToLengths
This will be used for converting assignPoints into values for storage by
EventResults.

I'm aiming to convert EventResults.indivSweeps, etc. into 2-D arrays
that can hold different sweeps values for each competitor. 1-D Fraction
arrays from Sweepstakes.assignPoints will be run through this method to
produce arrays that represent assigning the same amount of points to all
competitors in a given tie. The user can modify this behavior if need be
by editing the array (GUI for this to be implemented).