Add Event.maxPoints and getter
A brief glance at version 1 of this app reminded me that a max-points
mechanism is necessary. I have thus added it.

This will be used when computing points for a school. (I'll probably add
a new method to EventResults that calls computeTotalSweeps but enforces
this cap.) I'll be adding to the Event constructor as well; that's
probably going to be the next commit.