BSS: Enforce minimum dimension of 800x600
This is to prevent infinite looping as the font-shrinking logic tries to
shrink a full name into a 2-pixel space. It might compromise
compatibility with mobile computing or with extremely old systems, but
this is intended to be projected in a common area, so 800x600 is (in my
opinion) a sensible minimum. (If needed, the DisplayPanel can set a
scale factor that allows an 800x600 slide to occupy a 400x300
canvas...but that's getting into feature creep.)