DisplayPanel: Clarify intent
What I intend for DisplayPanel to be is a drawing surface that will be
controlled by a Selector (to be integrated into logic later). The
DisplayPanel will be added to a JFrame that will pass in a Selector to
the DisplayPanel constructor (again, to be implemented). The JFrame will
then manage all communication between the DisplayPanel and the outside
world. This allows me to make a UserSlideDisplayFrame (which takes user
input from Swing controls), a SlideTestFrame (which displays one Slide
consistently), a ProtocolDisplayFrame (which uses a ProtocolSelector for
the actual application behavior that will be used in projection), etc.