Add EditorTableModel
This will be used by ERED and the forthcoming SED
(SweepstakesEditorDialog) in order to allow the user to edit data.

----

WHY I HAVE BEEN INACTIVE
========================

First, I've been in the process of migrating to a new PC. I've had to
transfer all of my configuration, etc. to this new computer. Having this
project on Github simplified this somewhat, but it still took time to
transfer e.g. my Eclipse configuration.

Second, classes have started and I thus have less time in general.