ETM: Add doc note that stages shouldn't change column headers
The reason for the stage system is because of ERED. I'm planning on two
stages. The first allows the user to input event results. In preparing
for the second stage, the application generates sweepstakes points using
the event's specification. Then, in the second stage, these are exposed
to the user for editing.