Add RankingEditorTableModel
--the aforementioned superclass. This will handle functionality such as
handling ties and adding rows at the end for ties for last place. For
example, it will create checkboxes in the first column for specifying
whether a row is a nonterminal member of a tie. (Additional rows will be
added/removed based on these checkboxes.)

I should probably ponder what I'll let this class handle and what I'll
let subclasses handle before writing this code. This might take anywhere
from a few nanoseconds to a few minutes. (to a few days...)

I also need to think about what happens if someone unchecks a box above
a row containing data (or several rows containing data) at the end of
the table. The logical behavior is to obliterate that data, since it's
no longer relevant (it corresponds to a tie that we're not tracking),
but we know that [l]users do stupid things all the time. (I'm talking to
you, `rm -rf .*`. Don't ever, EVER type that on a *nix system, by the
way, no matter your PWD.) Optimally, the system will store that data
somehow (maybe just hide the rows instead of deleting them) or present a
confirmation box before deleting the data.