RETM: Add invisible (initialization NYI)
Consider a four-way tie for last place. If an operator accidentally
checks the "End tie?" box for the top member of the tie, all of the
other rows will extend beyond retainPlaces and be removed. Having an
invisible list allows this data to be easily recovered (having to
confirm all tie edits would be too much of a hassle).

We will store the value of the boolean "end tie" field (so that an
accidentally deleted tie can be reconstructed in one click). If we reach
the end of the stack and the last popped member has this field set to
false, a new row should be inserted at the end with default values.