Add RETM.updateForNewRowCount, excluding event firing
This will form the backbone of our logic and will be executed from other
methods. Since event firing doesn't have to do with the logic of adding
or removing rows, I'll be implementing it later (probably in the next
commit).

My unofficial goal for the RETM class is to handle all of the
functionality itself, with subclasses (e.g. those dedicated to editing
EventResults or sweeps) only specifying the contents and doing any
specific logic (e.g. a school autocomplete, time permitting).