RETM: Allow subclasses to recompute rows (NYI: event firing)
I'm hoping for the sweepstakes editor to pull the sweepstakes that would
be assigned based solely on the logic for the events and allow the user
to enter an "offset" from this value. The sum would then be dynamically
computed. This is an example of a use case for this method.

I should probably make some of these methods final.