RETM: Fire events when computing rows
CAVEAT: This could lead to infinite recursion. To solve this, I'll add a
check in the tableChanged method that we haven't already computed this
field. I might save all of the editable fields whenever we compute a row
and, before computing the row, check that we haven't already computed it
for the given values. (Of course, if subclassers mess with editable
fields when recomputing, they should take care not to recurse
infinitely.)

computeRows(1, 1)
-> computeRow(1)
-> fireTableRowsUpdated(1, 1)
(I'll add an event listener later)
-> tableChanged(relevant TableModelEvent)
(presumably)
-> computeRows(relevant TableModelEvent)
-> computeRows(1, 1)