Extract a version of RETM.vCC that operates on an array, not an index
This will make my class-validation logic more flexible.

As a side-note, I'll probably not pass the reference to the table row
directly into the abstract computeRow method. Instead, I'll copy the
array, pass it to the subclass, and call vCC before replacing the array.
Slower, but ensures that the data are always valid.